pyproject.toml
[tool.poetry]
name = "vespa-sample-app"
version = "0.1.0"
description = "search app using vespa"
authors = ["Kyohei-Prj <tatsukawaforjob@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyvespa = "^0.40.0"
pandas = "^2.2.2"
pyarrow = "^16.0.0"
fastparquet = "^2024.2.0"
sentence-transformers = "^2.7.0"
click = "^8.1.7"
loguru = "^0.7.2"
streamlit = "^1.33.0"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.360"
tree-sitter-languages = "^1.10.2"
autopep8 = "^2.1.0"
pytest = "^8.2.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
